dev_requirements.txt
Flask==0.10.1
PyYAML==3.11
coverage==3.7.1
fudge==1.0.3
mock==1.0.1
nose==1.3.3
simple-crypt==4.0.0
